+++ src/rollup.js
@@ -4,7 +4,7 @@ import { writeFile } from './utils/fs.js';
 import { assign, keys } from './utils/object.js';
 import { mapSequence } from './utils/promise.js';
 import validateKeys from './utils/validateKeys.js';
-import SOURCEMAPPING_URL from './utils/sourceMappingURL.js';
+import { SOURCEMAPPING_URL } from './utils/sourceMappingURL.js';
 import Bundle from './Bundle.js';
 
 export const VERSION = '<@VERSION@>';
@@ -67,7 +67,17 @@ export function rollup ( options ) {
 	return bundle.build().then( () => {
 		timeEnd( '--BUILD--' );
 
-		function generate ( options ) {
+		function generate ( options = {} ) {
+			if ( !options.format ) {
+				bundle.warn({
+					code: 'MISSING_FORMAT',
+					message: `No format option was supplied – defaulting to 'es'`,
+					url: `https://github.com/rollup/rollup/wiki/JavaScript-API#format`
+				});
+
+				options.format = 'es';
+			}
+
 			timeStart( '--GENERATE--' );
 
 			const rendered = bundle.render( options );